Remove specific functionality from InteractiveAssistantWithGUI class
This commit removes specific functionality from the InteractiveAssistantWithGUI class in the 'assistant_with_gui.py' file. The following functionality has been removed:

- Wikipedia search
- Opening and searching specific applications (YouTube, Google, StackOverflow, AISC, AISC forum)
- Definition lookup

The corresponding if-elif blocks have been  removed. The class still retains the ability to display user input on the GUI.

The commands will be rebuilded and re initialized in future commits.